// mlflow/server/js/src/shared/web-shared/genai-traces-table/types.ts
export type NumericAggregateCount = {
  // The lower bound of the bucket, inclusive.
  lower: number;
  // The upper bound of the bucket, exclusive except for the last bucket.
  upper: number;
  // The number of values in the bucket.
  count: number;
};

// A numeric aggregate with the min, mid, and max values, and the counts of values in each bucket.
export type NumericAggregate = {
  min: number;
  max: number;
  maxCount: number;
  counts: NumericAggregateCount[];
};